feat(web): transcribe node config form; drop dead START voice-model picker
Transcription moved from the START node to the dedicated transcribe node
on the backend; give it a bespoke config form (voiceModel picker +
saveAsUserMessage toggle) registered in the custom-widgets dispatch, and
remove the now-dead voice-model picker from START (acceptVoice stays).
