+++ assemblix-app-web/src/entities/workflow/lib/workflow-editor/ui/node-forms/index.ts
@@ -5,6 +5,7 @@ export { ConditionNodeForm } from "./condition-node-form";
 export { SetVariableNodeForm } from "./set-variable-node-form";
 export { EndNodeForm } from "./end-node-form";
 export { HTTPRequestNodeForm } from "./http-request-node-form";
+export { TranscribeNodeForm } from "./transcribe-node-form";
 export { useNodeDataChange } from "./useNodeDataChange";
 export { GenericNodeForm } from "./generic/generic-node-form";
 export { customWidgets } from "./generic/custom-widgets";
